Find cyclones information

Find cyclones information

Usage

find_bagyo(.year = NULL, .category = NULL)

Arguments

.year

An integer value for a year or a vector of years of cyclones data to retrieve. Default to NULL to retrieve all years.

.category

A character value or a vector of category code/s or category name/s to retrieve. Default to NULL to retrieve all categories

Value

A data.frame of cyclones information based on specified year and category

Examples

find_bagyo()
#> # A tibble: 119 × 9
#>     year category_code category_name         name  rsmc_name start              
#>    <dbl> <fct>         <fct>                 <chr> <chr>     <dttm>             
#>  1  2017 TD            Tropical Depression   Auri… NA        2017-01-07 08:00:00
#>  2  2017 TD            Tropical Depression   Bisi… NA        2017-02-03 14:00:00
#>  3  2017 TD            Tropical Depression   Cris… NA        2017-04-14 14:00:00
#>  4  2017 TS            Tropical Storm        Dante Muifa     2017-04-26 08:00:00
#>  5  2017 STS           Severe Tropical Storm Emong Nanmadol  2017-07-02 02:00:00
#>  6  2017 TD            Tropical Depression   Fabi… Roke      2017-07-22 02:00:00
#>  7  2017 TY            Typhoon               Gorio Nesat     2017-07-25 14:00:00
#>  8  2017 TS            Tropical Storm        Huan… Haitang   2017-07-30 02:00:00
#>  9  2017 STS           Severe Tropical Storm Isang Hato      2017-08-20 08:00:00
#> 10  2017 TS            Tropical Storm        Joli… Pakhar    2017-08-24 14:00:00
#> # ℹ 109 more rows
#> # ℹ 3 more variables: end <dttm>, pressure <int>, speed <int>
find_bagyo(.year = 2017)
#> # A tibble: 22 × 9
#>     year category_code category_name         name  rsmc_name start              
#>    <dbl> <fct>         <fct>                 <chr> <chr>     <dttm>             
#>  1  2017 TD            Tropical Depression   Auri… NA        2017-01-07 08:00:00
#>  2  2017 TD            Tropical Depression   Bisi… NA        2017-02-03 14:00:00
#>  3  2017 TD            Tropical Depression   Cris… NA        2017-04-14 14:00:00
#>  4  2017 TS            Tropical Storm        Dante Muifa     2017-04-26 08:00:00
#>  5  2017 STS           Severe Tropical Storm Emong Nanmadol  2017-07-02 02:00:00
#>  6  2017 TD            Tropical Depression   Fabi… Roke      2017-07-22 02:00:00
#>  7  2017 TY            Typhoon               Gorio Nesat     2017-07-25 14:00:00
#>  8  2017 TS            Tropical Storm        Huan… Haitang   2017-07-30 02:00:00
#>  9  2017 STS           Severe Tropical Storm Isang Hato      2017-08-20 08:00:00
#> 10  2017 TS            Tropical Storm        Joli… Pakhar    2017-08-24 14:00:00
#> # ℹ 12 more rows
#> # ℹ 3 more variables: end <dttm>, pressure <int>, speed <int>
find_bagyo(.category = "TD")
#> # A tibble: 32 × 9
#>     year category_code category_name       name    rsmc_name start              
#>    <dbl> <fct>         <fct>               <chr>   <chr>     <dttm>             
#>  1  2017 TD            Tropical Depression Auring  NA        2017-01-07 08:00:00
#>  2  2017 TD            Tropical Depression Bising  NA        2017-02-03 14:00:00
#>  3  2017 TD            Tropical Depression Crising NA        2017-04-14 14:00:00
#>  4  2017 TD            Tropical Depression Fabian  Roke      2017-07-22 02:00:00
#>  5  2017 TD            Tropical Depression Nando   NA        2017-09-23 14:00:00
#>  6  2018 TD            Tropical Depression Josie   NA        2018-07-20 19:00:00
#>  7  2018 TD            Tropical Depression Luis    NA        2018-08-22 18:00:00
#>  8  2018 TD            Tropical Depression Neneng  Barijat   2018-09-08 06:00:00
#>  9  2018 TD            Tropical Depression Usman   NA        2018-12-25 10:00:00
#> 10  2019 TD            Tropical Depression Amang   NA        2019-01-19 06:00:00
#> # ℹ 22 more rows
#> # ℹ 3 more variables: end <dttm>, pressure <int>, speed <int>
find_bagyo(.year = 2017, .category = "TD")
#> # A tibble: 5 × 9
#>    year category_code category_name       name    rsmc_name start              
#>   <dbl> <fct>         <fct>               <chr>   <chr>     <dttm>             
#> 1  2017 TD            Tropical Depression Auring  NA        2017-01-07 08:00:00
#> 2  2017 TD            Tropical Depression Bising  NA        2017-02-03 14:00:00
#> 3  2017 TD            Tropical Depression Crising NA        2017-04-14 14:00:00
#> 4  2017 TD            Tropical Depression Fabian  Roke      2017-07-22 02:00:00
#> 5  2017 TD            Tropical Depression Nando   NA        2017-09-23 14:00:00
#> # ℹ 3 more variables: end <dttm>, pressure <int>, speed <int>