Philippine Tropical Cyclones Data

The Philippines frequently experiences tropical cyclones (called 'bagyo' in the Filipino language) because of its geographical position. These cyclones typically bring heavy rainfall, leading to widespread flooding, as well as strong winds that cause significant damage to human life, crops, and property. Data on cyclones are collected and curated by the Philippine Atmospheric, Geophysical, and Astronomical Services Administration (PAGASA) and made available through its website. This package contains Philippine tropical cyclones data in a machine-readable format. It is hoped that this data package provides an interesting and unique dataset for data exploration and visualisation as an adjunct to the traditional iris dataset and to the current palmerpenguins dataset.

Usage

bagyo

Format

A data frame with 9 columns and 86 rows:

Variable	Description

year	Year

category_code	Tropical cyclone category code

category_name	Tropical cyclone category name

name	Name given to the tropical cyclone by Philippine authorities

rsmc_name	Name given to the tropical cyclone by RSMC

start	Date and time at which cyclone enters Philippine waters

end	Date and time at which cyclone leaves Philippine waters

pressure	Maximum central pressure in hPa

speed	Maximum sustained wind speed in km/h

Source

Data are drawn from PAGASA's Annual Report on Philippine Tropical Cyclones found at https://www.pagasa.dost.gov.ph/tropical-cyclone/publications/annual-report

See also

Author

Maintainer: Ernest Guevarra ernest@guevarra.io (ORCID) [copyright holder]

Examples

bagyo
#> # A tibble: 101 × 9
#>     year category_code category_name         name  rsmc_name start              
#>    <dbl> <fct>         <fct>                 <chr> <chr>     <dttm>             
#>  1  2017 TD            Tropical Depression   Auri… NA        2017-01-07 08:00:00
#>  2  2017 TD            Tropical Depression   Bisi… NA        2017-02-03 14:00:00
#>  3  2017 TD            Tropical Depression   Cris… NA        2017-04-14 14:00:00
#>  4  2017 TS            Tropical Storm        Dante Muifa     2017-04-26 08:00:00
#>  5  2017 STS           Severe Tropical Storm Emong Nanmadol  2017-07-02 02:00:00
#>  6  2017 TD            Tropical Depression   Fabi… Roke      2017-07-22 02:00:00
#>  7  2017 TY            Typhoon               Gorio Nesat     2017-07-25 14:00:00
#>  8  2017 TS            Tropical Storm        Huan… Haitang   2017-07-30 02:00:00
#>  9  2017 STS           Severe Tropical Storm Isang Hato      2017-08-20 08:00:00
#> 10  2017 TS            Tropical Storm        Joli… Pakhar    2017-08-24 14:00:00
#> # ℹ 91 more rows
#> # ℹ 3 more variables: end <dttm>, pressure <int>, speed <int>